[hijax] fix caching logic for newstyle hi primitives
The trouble was that we memoize abstract eval, as a function on the primitive,
input avals, and params, but with a newstyle hi primitive one might have
parameters that affect the output aval but that we don't save on self.params
because the out aval itself has them as attributes.

The fix is not to apply the usual abstract eval memoizing on newstyle hi
primitives, since newstyle hi primitives have the output avals as parameters.
We don't need a separate cache; we can just read the out avals instead!