[JAX] Require IFRT values and executables to have some `xla::ifrt::UserContext` if they are wrapped as JAX objects
JAX will check at runtime if IFRT values and executables have `xla::ifrt::UserContext` if they are wrapping them as JAX objects.

This replaces `-DIFRT_REQUIRE_USER_CONTEXT` for checking if IFRT user
context setups is done correctly.

Limitation: This check does not cover the paths that do not wrap any IFRT
values and executables, notably blocking on an array and fetching an
array to host.
PiperOrigin-RevId: 825661892