[NFC] Add `name` argument to `pallas_call` primitive.
This change introduces an optional `name` string argument to the `pallas_call` primitive and its associated lowering and transformation rules. The `name` is currently passed through but not yet used by the backend-specific lowering implementations for TPU, GPU, and Triton.

A follow up CL will use it to extend the name stack on TPU

PiperOrigin-RevId: 826174027