[Mosaic GPU] Fix lowering of `InitializeBarrierOp`.
In [#30545](https://github.com/jax-ml/jax/pull/30545), we replaced the conversion from LLVM pointer to Memref with an unrealized conversion cast. We forgot to update the lowering of `InitializeBarrierOp` accordingly.

The reason this is not a problem is because the result of `InitializeBarrierOp` is never used. We could drop the returned value from the op but this is orthogonal to this change.

PiperOrigin-RevId: 826498254