[Mosaic GPU][NFC] Consolidate `is_signed` derivation in dialect lowering.
When lowering from Pallas dtype to IR type, we loose the `is_signed` information. IR uses unsigned integers for both signed and unsigned integers. When lowering MGPU operations, this information is either inferrable from the op (e.g. `arith.MaxSIOp`), irrelevant (e.g. `vector.LoadOp`) or needs to be recovered (e.g. `mgpu.WGMMAOp`). We propose to annotate the op in Pallas with an attribute when the information is relevant.

PiperOrigin-RevId: 827464618