Replace `set_xla_metadata` to `set_xla_debug_metadata` and deprecate the former
The `set_xla_metadata` mechanism is only good for debugging purpose because it tries to preserve the computation graph structure at the cost of causing cache misses and retracing, which can be slow. So this renaming makes the intention clear and hopefully can encourage people to only use it for debugging purpose and nothing else.

PiperOrigin-RevId: 828126517