[Pallas/Mosaic GPU] Fix `emit_pipeline` to handle dynamic grids containing less than `max_concurrent_steps` iterations.
Previously, we would fail to reduce `max_concurrent_steps` (unlike in the
static case), which would lead to us arriving too many times on our barriers,
and therefore to a `cuLaunchKernel` failure.

The change includes a regression test. Its setup is fairly contrived, but I
didn't find a better way to reproduce the issue in a simple test.

PiperOrigin-RevId: 829375557