[Pallas/Mosaic TPU] Allow non-leading and non-matching batch dimensions in `dot_general`.
The constraints on `lhs_batch_dims` and `rhs_batch_dims` for `dot_general` in Pallas/Mosaic on TPU are now relaxed. Batch dimensions do not have to be at the front of the shape, and the dimension indices used for batching on the LHS and RHS can be different.

The remaining gap compared to JAX is the lack of support for multiple batch dimensions.

PiperOrigin-RevId: 831015967