[Mosaic:TPU][infer-vector-layout] Fix bug in tpu.concatenate rule where invalid offsets can be assigned
Sometimes we would preserve the input layout offsets even when the input tiling changed. For example, for a 1D concat, if one shape has layout {4,0},(8, 128),-2 in its producer, it would be assigned a layout of {4,0},(1,128),-2, which is invalid.

Now, we correct the offset to fit in the vreg slice for the chosen tiling.

PiperOrigin-RevId: 831175068