[Mosaic GPU] Optimize i4 -> i8 conversion by packing registers if vec_len < 8.
This is done with care to not undo the optimization where we recognize the registers
already originate from slices of 32-bit regs and can avoid the concatenation ops altogether.
The benchmarks indicate that WGMMA_LAYOUT (1591 -> 1303 instrs) and WGMMA_LAYOUT_UPCAST_2X
(631 -> 471 instrs) benefit from the change, while all other cases remain with the same
instruction counts as before the change.

PiperOrigin-RevId: 831858710