[MLIR] Fix construction of `i1` attributes from numpy arrays and add tests.
This change fixes the construction of MLIR attributes from `numpy` arrays for a Boolean `dtype`. This happened when calling `jax.mlir.ir_attribute` and the problem occurred in `_numpy_array_attribute`: that function *first* bit-packed the Booleans and *then* determined the `dtype` from the result, which is `uint8` at that point rather than `bool`. The fix thus consists of a simple swapping of lines. Note that the virtually equivalent `_numpy_array_constant` already used the right order.

The change extends and improves the tests for the FFI primitive, which uses `jax.mlir.ir_attribute` under the hood. The previous tests use a flawed comparison to verify the round-tripped type: it checked whether the *Python* type of the created MLIR attribute is the expected one. However, that is not enough: two `DenseElementAttr`s with different element types pass that check but really should not. That check would not have caught the bug fixed in this change (plus it didn't test the variant with `bool`s). The new version uses strings instead, which is not only more precise but also more concise in building the expected values. In addition to this improvement, the change adds instances of the corresponding FFI test for essentially all attribute types I could find.

PiperOrigin-RevId: 834770477