[Mosaic GPU] Implement `untile_shape`, a MLIR operation that allows untiling a memref.
This op is scaffolding for applying BlockSpec transforms in upcoming Pallas changes.

Essentially, tracing forces us to produce references with their type already
transformed in the case when transforms are specified in BlockSpecs. As a
result, these refs can't be handled directly when lowering with warpgroup
semantics, since we expect refs to be untransformed in that case. This op will
essentially allow us to undo a `TileTransform` at lowering time, resolving the
mismatch.

Once we get rid of lane semantics lowering, the hope is that this op can just be deleted.

PiperOrigin-RevId: 837772095