[export] Add support for explicit sharding.
This is a significant addition to exporting JAX functions.

Previously, the `Exported` objects and their serializations
did not store the meshes, in the idea that one may want to
use the `Exported` object with another concrete mesh, as long
as the number of devices match.

Starting with this PR, we added new fields to `Exported`
to contain NamedShardings (including abstract meshes, and
partition specs). This is needed to suport explicit
sharding.

One added advantage of this change is that we do not
need to rely on Shardy to store the mesh in the serialized
HLO. (Those Shardy-stored meshes do not support multiple
meshes for different inputs/ouputs, and currently do not
include the `axis_type`).

There is one regression: The function `in_shardings_jax(new_mesh)`
previously cared only for `new_mesh` to have the same number
of devices. Now we check that the `new_mesh` has the same
abstract mesh as the one that was used for serialization.