`jax.nn.standardize` fix NaN output.
Fixes #30426. The root of the issue is that we use a performant but inaccurate approach to computing the variance, which results in near-zero variance sometimes being negative. There are a few ways we could address this:

1. Use a more accurate variance computation. I did not explore this, because I suspect it would have performance impact on downstream code.
2. Use `maximum(var, eps)` rather than `var + eps`. I did not opt for this, because it could change numerics and affect autodiff for variances near epsilon.
3. Use `abs(var) + eps`. This approach should be fine, because in cases of interest the variance will be vanishingly small, and so gradients will not be affected.
4. Use `clip(var, 0) + eps`. This approach should also be fine for the same reason as 3.

Opting here for (4) & running a suite of downstream tests shows no failures with this approach.

PiperOrigin-RevId: 838825157