Make every parallel primitive multiple_results=False. Primitives include `psum_p`, `psum_invariant_p`, `pmax_p`, `pmin_p`, `pvary_p`, `preduced_p`, `unreduced_psum_p`, `vary_unreduced_cast_p` and `reduced_vary_cast_p`
This lead to massive simplification of the code.

Some history: `psum_p` had `multiple_results=True` because a long long time ago, psum_p emitted a single all-reduce for multiple axes. https://github.com/jax-ml/jax/pull/2337 was the PR.

But after the introduced of stableHLO lowerings, psum_p always emitted multiple all-reduces instead of a single all-reduce and it'll be a pain to change that now. So let's simplify the code.

We can bring up multiple_results=True in the future if we want to emit a single all-reduce. So this change does not get rid of that possibility.

PiperOrigin-RevId: 839323775