[export] Add backwards compatibility tests for serialized exports
Recently we have found the need to evolve the serialization of
`jax.export.Exported`. E.g., in #33942 we have added a 32-bit
representation for `nr_devices`. This introduced a compatiblity
bug that was found by usersm and fixed in in #33685.

Here we add backwards compatibility tests. See the description
in the `export_serialization_back_compat_test.py` module docstring.

Note that this is separate from our previous set of backwards
compatibility tests for the lowering of custom calls
(in `export_back_compat_test.py`). However, we reuse some
of the same ideas, and we use the same directory for
storing saved old serializations.