[Mosaic GPU] Get rid of hints in layout inference.
All the hints we derive can now be constructed on the fly based on `Relayout`s
or `Equals(Reduce())` constraints. This greatly simplifies the necessary
bookkeeping in the layout inference pass, as well as the APIs.

A follow-up change will take care of deleting support for
`{Least,Most}ReplicatedExpression`, which no longer need to exist.

PiperOrigin-RevId: 841933775