[hijax] handle symbolic zeros generically in hijax lin rule
Soon we'll upgrade the user-facing rule api to support symbolic zeros. That's
in the custom-vjp3 branch.

But to unblock things now, and as a convenience wrapper later, we can support
not differentiating with respect to some inputs without the user rule being
aware: we just use the black-hole /dev/null NullAccum in the right places,
where the right places are indicated by the in_nz on the linearize pass. This
is less good than telling the user rule about in_nz because it doesn't let the
user rule save any work or memory.

(At first we tried just instantiating zeros, but that was worse because we lose
information as to whether those zeros are linear or nonlinear, complicating
things downstream.)

Co-authored-by: Robert Dyro <rdyro@google.com>