[mutable-arrays] ignroe InternalMutableArrayEffect in partial_eval has_effects
It's not really an effect from the point of view of partial eval; like
NamedAxisEffect, it's just a bit of info we're sneaking into the effect system
as a handy way of tracking it. But it should be ignored for the purpose of
deciding "is this code effectful".

(We're hoping that InternalMutableArrayEffect, and the effect system more
generally, and partial eval, are not long for this world...)

This is a follow-up on #33906

Co-authored-by: Sharad Vikram  <sharadmv@google.com>