[JAX] Track the layout defaultness more precisely for arrays created using `from_dlpack`
This change checks if a JAX array was created using an input buffer copy in `from_dlpack`. The copying would make the newly created `PjRtBuffer` to use a default layout. This defaultness would be tracked correctly.

This change is expected to be a no-op today (and does not cause a problem before this change, either). Previously, the array with the new `PjRtBuffer` with copying would take the concrete default layout as if it were a custom layout, and now the same concrete layout will be available through `array.layout`. Internally in the runtime, it would now know that the defaultness, and knowing this aspect will be required in subsequent IFRT layout support changes.

PiperOrigin-RevId: 845946765