[Mosaic GPU] Add a sanity check ensuring that layout inference only ever assigns layouts compatible with variables.
Previously, when no such check existed, we would sometimes allow assigning
layouts incompatible with the shape of a variable, especially in the case of
layout casts.

As a result of this change, a few tests that were assigning invalid layouts
had to be updated, as well as the rules that perform variable assignments,
i.e. layout casts in different memory spaces, and MMA ops.

To propagate the errors, we allow the derivation rules to return
`Unsatisfiable` directly. Later, we may want to think about improving error
reporting, since this should allow for more helpful diagnostics.

PiperOrigin-RevId: 846722181